move all parsing of ws messages to own module
websocket module was starting to get very messy with parsing stuff, this cleans that up a bit
