+++ twitch/__init__.py
@@ -7,7 +7,7 @@ __copyright__ = 'Copyright 2019-2019 sedruk'
 __version__ = '0.1.0'
 
 from .client import Client
-from .event import Event
 from .user import User
 from .message import Message
 from .channel import Channel
+from .events import Event
