+++ twitch/event_handler.py
@@ -1,45 +1,8 @@
 import logging
 import asyncio
-import collections
+from .events import Event
 
 log = logging.getLogger(__name__)
-
-
-class Event:
-    # heartbeat/keep-alive events
-    PINGED = 'ping'
-    PONGED = 'pong'
-
-    # request events
-    TAG_REQUEST_ACKED = 'tag_request_acked'
-
-    # connection state events
-    CONNECTED = 'connected'
-    DISCONNECT = 'disconnect'
-    AUTHENTICATED = 'authenticated'
-
-    # raw socket events
-    SOCKET_SEND = 'socket_send'
-    SOCKET_RECEIVE = 'socket_receive'
-
-    # standard events
-    MESSAGE = 'message'
-    USER_JOIN_CHANNEL = 'user_join_channel'
-    USER_LEFT_CHANNEL = 'user_leave_channel'
-    LIST_CHATTERS = 'list_chatters'
-    MOD_STATUS_CHANGED = 'mod_status_updated'
-    CHAT_CLEARED = 'chat_cleared'
-    MESSAGE_CLEARED = 'message_cleared'
-    HOST_MODE_CHANGED = 'host_mode_updated'
-    CHANNELS_REJOINED = 'channels_rejoined'
-    CHANNEL_STATE_CHANGED = 'channel_state_update'
-    USER_NOTIFICATION = 'user_noitification'
-    USER_STATE_CHANGED = 'user_updated'
-
-    # unknown opcode
-    # will notify listeners about any undocumented and/or
-    # error messages from twitch
-    UNKNOWN = 'unknown'
 
 
 class _EventTask(asyncio.Task):
