+++ twitch/http.py
@@ -8,7 +8,8 @@ from weakref import WeakValueDictionary
 import aiohttp
 
 from . import __version__
-from .exception import TwitchException
+from .exception import HTTPException, HTTPNotAuthorized, HTTPNotFound, \
+    HTTPForbidden
 
 log = logging.getLogger(__name__)
 
@@ -27,37 +28,6 @@ class LockHelper:
     def __exit__(self, type, value, traceback):
         if self._unlock:
             self.lock.release()
-
-
-class HTTPException(TwitchException):
-    def __init__(self, response, error, status=None):
-        self.status = response.status if not status else status
-        self.response = response
-        self.message = None
-        if isinstance(error, dict):
-            self.message = error.get('message', 'Unknown')
-            self.error = error.get('error', '')
-        else:
-            self.error = error
-
-        reason = f' (reason: {self.message})' if self.message else ''
-        e = f'{self.status} {self.error}{reason}'
-        super().__init__(e)
-
-
-class HTTPNotAuthorized(HTTPException):
-    def __init__(self, response, error):
-        super().__init__(response, error, 401)
-
-
-class HTTPForbidden(HTTPException):
-    def __init__(self, response, error):
-        super().__init__(response, error, 403)
-
-
-class HTTPNotFound(HTTPException):
-    def __init__(self, response, error):
-        super().__init__(response, error, 404)
 
 
 class HTTPRoute:
